refactor: extract logger from HopContextImpl
